Issue-457: Moved UIGenre to core module
