--- app/src/main/java/com/kirchhoff/movies/data/ui/core/UIGenre.kt
@@ -1,3 +1,0 @@
-package com.kirchhoff.movies.data.ui.core
-
-data class UIGenre(val name: String)
