+++ app/src/main/java/com/kirchhoff/movies/mapper/core/ICoreMapper.kt
@@ -1,7 +1,7 @@
 package com.kirchhoff.movies.mapper.core
 
+import com.kirchhoff.movies.core.data.UIGenre
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
-import com.kirchhoff.movies.data.ui.core.UIGenre
 import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentCredits
 import com.kirchhoff.movies.networkdata.core.NetworkObjectWithName
