+++ app/src/test/java/com/kirchhoff/movies/mapper/core/CoreMapperTest.kt
@@ -1,8 +1,8 @@
 package com.kirchhoff.movies.mapper.core
 
+import com.kirchhoff.movies.core.data.UIGenre
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentCredits
 import com.kirchhoff.movies.data.ui.core.UIEntertainmentPerson
-import com.kirchhoff.movies.data.ui.core.UIGenre
 import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentCredits
 import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentPerson
 import com.kirchhoff.movies.networkdata.core.NetworkObjectWithName
