+++ core/src/main/java/com/kirchhoff/movies/core/data/UIGenre.kt
@@ -1,0 +1,3 @@
+package com.kirchhoff.movies.core.data
+
+data class UIGenre(val name: String)
